pyproject.toml
[tool.poetry]
name = "fastapi-tortoise-orm"
version = "0.1.0"
description = "Ecommerce app using Tortoise ORM"
authors = ["Ckam <ckam225@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
tortoise-orm = "^0.18.1"
asyncpg = "^0.25.0"
aerich = "^0.6.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.9.0"}
httpx = "^0.22.0"
celery = "^5.2.3"
redis = "^4.1.4"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"
pytest = "^7.0.1"
requests = "^2.27.1"
pytest-asyncio = "^0.18.1"

[tool.aerich]
tortoise_orm = "core.settings.TORTOISE_ORM"
location = "database/migrations"
src_folder = "./."

[tool.pytest.ini_options]
asyncio_mode='auto'
addopts = "-ra -q"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
